refactor(finances): extract FinancesTab helpers to separate module
Pure functions, type interfaces, and constants (facilities, sponsor,
board-support) moved from FinancesTab.tsx into FinancesTab.helpers.ts.
The component file drops ~185 lines and is now focused on state and JSX.
