+++ src/components/finances/FinancesTab.tsx
@@ -1,12 +1,7 @@
 import { useEffect, useState } from "react";
 import { invoke } from "@tauri-apps/api/core";
 import { FinanceCashFlowChart } from "./FinanceCashFlowChart";
-import {
-  GameStateData,
-  MessageAction,
-  MessageData,
-  PlayerSelectionOptions,
-} from "../../store/gameStore";
+import { GameStateData, PlayerSelectionOptions } from "../../store/gameStore";
 import { Card, CardHeader, CardBody, Badge, ProgressBar, Button, Checkbox } from "../ui";
 import { User } from "lucide-react";
 import {
@@ -22,199 +17,32 @@ import {
   annualAmountToWeeklyCommitment,
   getTeamFinanceSnapshot,
 } from "../../lib/finance";
-import {
-  getFinanceSnapshot,
-  type FinanceSnapshotData,
-  type TeamFinanceSnapshotData,
-} from "../../services/financeService";
+import { getFinanceSnapshot } from "../../services/financeService";
 import { useTranslation } from "react-i18next";
 import ContextMenu from "../ContextMenu";
 import { translatePositionAbbreviation } from "../squad/SquadTab.helpers";
 import { resolveBackendError, resolveMessage } from "../../utils/backendI18n";
-
-type FacilityId = "Training" | "Medical" | "Scouting";
-
-interface FacilityUpgradeErrorState {
-  facilityId: FacilityId;
-  message: string;
-}
-
-interface FacilityDefinition {
-  effectKey: string;
-  id: FacilityId;
-  levelKey: "training" | "medical" | "scouting";
-  titleKey: string;
-}
-
-const DEFAULT_FACILITIES = {
-  training: 1,
-  medical: 1,
-  scouting: 1,
-};
-
-const FACILITY_DEFINITIONS: FacilityDefinition[] = [
-  {
-    id: "Training",
-    levelKey: "training",
-    titleKey: "finances.facilityTraining",
-    effectKey: "finances.facilityTrainingEffect",
-  },
-  {
-    id: "Medical",
-    levelKey: "medical",
-    titleKey: "finances.facilityMedical",
-    effectKey: "finances.facilityMedicalEffect",
-  },
-  {
-    id: "Scouting",
-    levelKey: "scouting",
-    titleKey: "finances.facilityScouting",
-    effectKey: "finances.facilityScoutingEffect",
-  },
-];
-
-function getFacilityUpgradeCost(level: number): number {
-  return level * 250_000;
-}
-
-function formatSignedAmount(value: number): string {
-  const formatted = formatVal(Math.abs(value));
-  return value < 0 ? `-${formatted}` : formatted;
-}
-
-function facilityUpgradeBlockReason(
-  snapshot: TeamFinanceSnapshotData,
-): string | null {
-  if (snapshot.currentlyOverBudget) {
-    return "be.error.finance.facilityUpgradeOverBudget";
-  }
-
-  if (
-    snapshot.overallStatus === "warning" ||
-    snapshot.overallStatus === "critical"
-  ) {
-    return "be.error.finance.facilityUpgradeCritical";
-  }
-
-  return null;
-}
-
-function boardSupportAvailable(snapshot: TeamFinanceSnapshotData): boolean {
-  return (
-    snapshot.currentlyInDebt ||
-    snapshot.runwayStatus === "warning" ||
-    snapshot.runwayStatus === "critical"
-  );
-}
-
-function sponsorPitchAvailable(snapshot: TeamFinanceSnapshotData): boolean {
-  return (
-    snapshot.currentlyOverBudget ||
-    snapshot.currentlyInDebt ||
-    snapshot.wageBudgetStatus === "warning" ||
-    snapshot.wageBudgetStatus === "critical" ||
-    snapshot.runwayStatus === "warning" ||
-    snapshot.runwayStatus === "critical"
-  );
-}
-
-function marketingCampaignAvailable(snapshot: TeamFinanceSnapshotData): boolean {
-  return sponsorPitchAvailable(snapshot);
-}
-
-function mapLocalFinanceSnapshot(
-  team: GameStateData["teams"][number],
-  snapshot: ReturnType<typeof getTeamFinanceSnapshot>,
-): TeamFinanceSnapshotData {
-  return {
-    annualWageBill: snapshot.annualWageBill,
-    weeklyWageSpend: snapshot.weeklyWageSpend,
-    weeklyWageBudget: snapshot.weeklyWageBudget,
-    weeklyRecurringIncome: snapshot.weeklySponsorIncome,
-    weeklySponsorIncome: snapshot.weeklySponsorIncome,
-    projectedWeeklyNet: snapshot.projectedWeeklyNet,
-    cashRunwayWeeks: snapshot.cashRunwayWeeks,
-    wageBudgetUsagePercent: snapshot.wageBudgetUsagePercent,
-    currentlyInDebt: team.finance < 0,
-    currentlyOverBudget: snapshot.annualWageBill > team.wage_budget,
-    wageBudgetStatus: snapshot.wageBudgetStatus,
-    runwayStatus: snapshot.runwayStatus,
-    overallStatus: snapshot.overallStatus,
-    marketingCampaignCooldownDaysRemaining:
-      snapshot.marketingCampaignCooldownDaysRemaining,
-  };
-}
-
-interface ResolveMessageActionResult {
-  game: GameStateData;
-  effect: string | null;
-  effect_i18n_key?: string | null;
-  effect_i18n_params?: Record<string, string> | null;
-}
-
-interface DelegatedRenewalResponseData {
-  game: GameStateData;
-  report: {
-    success_count: number;
-    failure_count: number;
-    stalled_count: number;
-  };
-}
-
-interface TaggedFinanceSnapshotData {
-  key: string;
-  data: FinanceSnapshotData;
-}
-
-interface BoardSupportResponseData {
-  game: GameStateData;
-  result: {
-    support_amount: number;
-    transfer_budget_reduction: number;
-    satisfaction_penalty: number;
-  };
-}
-
-interface SponsorPitchResponseData {
-  game: GameStateData;
-  result: {
-    message_id: string;
-    sponsor_name: string;
-    weekly_amount: number;
-    duration_weeks: number;
-  };
-}
-
-interface MarketingCampaignResponseData {
-  game: GameStateData;
-  result: {
-    message_id: string;
-    gross_revenue: number;
-    campaign_cost: number;
-    net_income: number;
-    cooldown_days: number;
-  };
-}
-
-function isChooseOptionAction(
-  actionType: MessageAction["action_type"],
-): actionType is {
-  ChooseOption: {
-    options: Array<{ id: string; label: string; description: string }>;
-  };
-} {
-  return typeof actionType === "object" && "ChooseOption" in actionType;
-}
-
-function isPendingSponsorOffer(message: MessageData): boolean {
-  return (
-    message.id.startsWith("sponsor_") &&
-    message.category === "Finance" &&
-    message.actions.some(
-      (action) => !action.resolved && isChooseOptionAction(action.action_type),
-    )
-  );
-}
+import {
+  type FacilityId,
+  type FacilityUpgradeErrorState,
+  type TaggedFinanceSnapshotData,
+  type ResolveMessageActionResult,
+  type DelegatedRenewalResponseData,
+  type BoardSupportResponseData,
+  type SponsorPitchResponseData,
+  type MarketingCampaignResponseData,
+  DEFAULT_FACILITIES,
+  FACILITY_DEFINITIONS,
+  getFacilityUpgradeCost,
+  formatSignedAmount,
+  facilityUpgradeBlockReason,
+  boardSupportAvailable,
+  sponsorPitchAvailable,
+  marketingCampaignAvailable,
+  mapLocalFinanceSnapshot,
+  isChooseOptionAction,
+  isPendingSponsorOffer,
+} from "./FinancesTab.helpers";
 
 interface FinancesTabProps {
   gameState: GameStateData;
